+++ tests/test_plot/test_int_bom.py
@@ -1262,6 +1262,18 @@ def test_int_bom_variant_rename_1(test_dir):
     ctx.clean_up()
 
 
+def test_int_bom_variant_rename_2(test_dir):
+    prj = 'f_rename_2'
+    ctx = context.TestContextSCH(test_dir, 'test_int_bom_variant_rename_2', prj, 'int_bom_var_rename_2_csv', BOM_DIR)
+    ctx.run(extra_debug=True)
+    rows, header, info = ctx.load_csv(prj+'-bom_(PROD).csv')
+    ref_column = header.index(REF_COLUMN_NAME)
+    check_kibom_test_netlist(rows, ref_column, 2, ['R2', 'D2'], ['R1', 'D1'])
+    rows, header, info = ctx.load_csv(prj+'-bom_(DEV).csv')
+    check_kibom_test_netlist(rows, ref_column, 3, None, ['R1', 'R2', 'D1', 'D2'])
+    ctx.clean_up()
+
+
 def test_int_bom_variant_t2s(test_dir):
     prj = 'kibom-variant_2'
     ctx = context.TestContextSCH(test_dir, 'test_int_bom_variant_t2s', prj, 'int_bom_var_t2s_csv', BOM_DIR)